chore: reorganize tree/treeseq code
