+++ src/trees/mod.rs
@@ -1,0 +1,5 @@
+mod tree;
+mod treeseq;
+
+pub use tree::Tree;
+pub use treeseq::TreeSequence;
